autoinstall: process params
Currently the autoinstall registration flags belongs to vdsm-reg
which should be deprecated. This patch adds autoinstall script
to manage the below kernel args:

- management_server
- management_server_port
- rhevm_admin_password
- ovirt_vdsm_disable

This patch will only parse the kernel arguments on the kernel
command line, it does not trigger the registration itself.
This is left to the generic registration once it lands.

Change-Id: Ibf664eb61344067427e4d88617dbbb7e8dd1d517
Signed-off-by: Douglas Schilling Landgraf <dougsland@redhat.com>