engine_page: Catch exception KeyError
If VDSM is not ready to provide network data via vdscli.getVdsCaps()['info']['networks']
and fail with KeyError we should throw an error message and not stuck in the login.
Additionally, for unexpected exceptions in vdscli we should log as error and not use
raise to avoid users getting blocked into the login page.

Change-Id: I4b7666c2020307c6119f96a2935684bc268e459b
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1198029
Signed-off-by: Douglas Schilling Landgraf <dougsland@redhat.com>