systemd: keep trying to register in case of failure during start
The current approach for autoinstall/autoregister is try
to register as soon supervdsm and network are ready but
in case it failure a new registration must be triggered manually.
This patch will make ovirt-node-plugin-vdsm service restart
each 60 seconds in case the service/registration couldn't be
triggered properly during the boot time. This behaviour is a
mimic from vdsm-reg which must be kept.

Change-Id: I8a50e5c8235413d428e9da8ceeec90e017eb63ee
Signed-off-by: Douglas Schilling Landgraf <dougsland@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/44063
Continuous-Integration: Jenkins CI
Reviewed-by: Fabian Deutsch <fabiand@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>